pyproject.toml
[project]
name="gravity-sim"
version="1.0.0"
description="A package to create simple 2D gravity simulations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "ruff>=0.13.0",
]

[project.scripts]
gravity-sim = "gravity_sim:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
